Refactor project structure and remove unused components
