+++ src/infrastructure/http/httpClient.ts
@@ -1,0 +1,2 @@
+import axios from 'axios';
+export const http = axios.create({ baseURL: import.meta.env.VITE_API_URL, withCredentials:true });
