--- src/views/AboutView.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="text-gray-200 dark:text-gray-400 size-100% ">
-    <h1>This is an about page</h1>
-  </div>
-</template>
